cloud_function/requirements.txt
google-cloud-storage==3.1.0
google-cloud-firestore==2.20.2
scikit-learn==1.6.1
joblib==1.5.0
numpy==2.2.5
